Refactor MemberDivision Methods
- Make the basic license details more generic so they can be reused
- Lift GetMemberDivisionAsync up to interface
- Add documentation comment
- Move ChunkInfo and EventType to Common for reuse
