+++ src/Aydsko.iRacingData/Common/EventType.cs
@@ -1,7 +1,7 @@
 ﻿// © 2022 Adrian Clark
 // This file is licensed to you under the MIT license.
 
-namespace Aydsko.iRacingData.Results;
+namespace Aydsko.iRacingData.Common;
 
 public enum EventType
 {
