+++ src/Aydsko.iRacingData/Common/License.cs
@@ -1,0 +1,25 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Text;
+using System.Threading.Tasks;
+
+namespace Aydsko.iRacingData.Common;
+
+public class License
+{
+    [JsonPropertyName("category_id")]
+    public int CategoryId { get; set; }
+    [JsonPropertyName("category")]
+    public string Category { get; set; } = default!;
+    [JsonPropertyName("license_level")]
+    public int LicenseLevel { get; set; }
+    [JsonPropertyName("safety_rating")]
+    public float SafetyRating { get; set; }
+    [JsonPropertyName("color")]
+    public string Color { get; set; } = default!;
+    [JsonPropertyName("group_name")]
+    public string GroupName { get; set; } = default!;
+    [JsonPropertyName("group_id")]
+    public int GroupId { get; set; }
+}
