+++ src/Aydsko.iRacingData/GlobalUsings.cs
@@ -1,5 +1,6 @@
 ﻿// © 2022 Adrian Clark
 // This file is licensed to you under the MIT license.
 
-global using Microsoft.Extensions.Logging;
 global using Aydsko.iRacingData.Common;
+global using Microsoft.Extensions.Logging;
+global using System.Text.Json.Serialization;
