+++ src/Aydsko.iRacingData/Stats/MemberDivision.cs
@@ -1,4 +1,6 @@
-﻿using Aydsko.iRacingData.Results;
+﻿// © 2022 Adrian Clark
+// This file is licensed to you under the MIT license.
+
 using System.Text.Json.Serialization;
 
 namespace Aydsko.iRacingData.Stats;
